Fix faster notification of completed bursts for IP
Only implemented for IP! WLAN Completely broken.

The idea is to check all flows every time we have an incoming packet
if they are to complete a burst. To make this almost scalable, we add
a FIFO queue with the time of added packets in flows, which can be
used to only check the relevant flows (like Dijkstras).

If a flow was added to the queue at time t, it then revceived a packet
at time t. Then, at time t', check and dequeue all items in the queue
enqueued before t' - inactive_time. If any of those flows had not had
new packets, send them to the output tx.